pyproject.toml
[tool.poetry]
name = "ddd-in-practice"
version = "0.1.0"
description = ""
authors = ["Felipe <felipe.melo@luizalabs.com>"]
readme = "README.md"
packages = [{include = "snack_machine"}, {include = "shared_kernel"}, {include= "management"}]

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.0.0"
sqlalchemy = "^2.0.19"
pika = "^1.3.2"
fastapi = "^0.100.1"
jinja2 = "^3.1.2"
uvicorn = "^0.23.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
rope = "^1.9.0"
pytest-cov = "^4.1.0"
httpx = "^0.24.1"
pylint = "^2.17.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
